Seperating DOM Manipulation frm Application Logic
