+++ src/modules/domVariables.js
@@ -1,0 +1,19 @@
+export const addTask = document.querySelector('.add-task');
+export const formTitle = document.querySelector('.form-title');
+export const editTask = document.querySelector('#update-task');
+export const sortItem = document.querySelector('#sort-item');
+export const submitGrpForm = document.querySelector('.add-group');
+export const deleteGroup = document.querySelector('#delete-group');
+export const submitTask = document.querySelector('#submit-task');
+export const taskBoxEdit = document.querySelector('.task-box');
+export const addForm = document.querySelector('#enter-task');
+export const groupName = document.querySelector('#group-name');
+export const groupBox = document.querySelector('#group-container');
+export const groupDupError = document.querySelector('#gErrMsg');
+export const taskName = document.querySelector('#task-name');
+export const dueDate = document.querySelector('#due-date');
+export const selectPriority = document.querySelector('#select-priority');
+export const describeTask = document.querySelector('#describe-task');
+export const taskEdit = document.querySelector('.fa-edit');
+export const taskDelete = document.querySelector('#del-task');
+export const taskDiv = document.querySelector('.task-box-div');
